+++ app/src/main/java/com/baidu/iov/dueros/waimai/net/entity/response/OrderListExtraBean.java
@@ -102,7 +102,7 @@ public class OrderListExtraBean {
             private double original_price;
             private int count;
             private String spec;
-            private int box_num;
+            private double box_num;
             private double box_price;
             private List<String> attrIds;
             private List<String> attrValues;
@@ -163,11 +163,11 @@ public class OrderListExtraBean {
                 return spec;
             }
 
-            public void setBox_num(int box_num) {
+            public void setBox_num(double box_num) {
                 this.box_num = box_num;
             }
 
-            public int getBox_num() {
+            public double getBox_num() {
                 return box_num;
             }
